requirements.txt
albumentations==1.0.3
opencv_python_headless==4.5.3.56
onnxruntime_gpu==1.8.1
numpy==1.21.1
onnxruntime==1.9.0
Pillow==8.4.0
PyQt5==5.15.4
